+++ packages/cli-plugin-package/src/index.ts
@@ -22,6 +22,7 @@ import {
   lstat,
   readFile,
   createWriteStream,
+  writeFile,
 } from 'fs-extra';
 import * as ts from 'typescript';
 import * as globby from 'globby';
@@ -66,6 +67,8 @@ export class PackagePlugin extends BasePlugin {
 
   private compilerHost: CompilerHost;
   private program: Program;
+  // 在 bundle 的情况下是否使用 hcc 编译 @midwayjs/hooks
+  private isUseHcc = false;
 
   commands = {
     package: {
@@ -689,27 +692,27 @@ export class PackagePlugin extends BasePlugin {
 
     const midwayBuildPath = this.core.config.buildPath;
     const distDir = join(this.midwayBuildPath, 'dist');
-    let isUseHcc = false;
     console.log('midwayBuildPath', midwayBuildPath);
     try {
       const modules = [];
-      // TODO: hooks bundle
       const { getPreloadCode } = require('@midwayjs/hcc');
       const hooksBundle = await getPreloadCode(midwayBuildPath, file => {
-        if (/midway_preload/.test(file)) {
+        if (/midway_preload|configuration/.test(file)) {
           return false;
         }
         modules.push(`require('./${file}')`);
         return true;
       });
 
-      const code = `${hooksBundle};exports.modules = [${modules.join(', ')}];`;
-
+      let code = `${hooksBundle};exports.modules = [${modules.join(', ')}];`;
+      // TODO: fix hooks bug
+      code = code.replace('will be overwritten.', 'will be overwritten.\n');
+      code = code.replace("require('./configuration.js')", '');
       console.log('hooksBundle code:', code);
-      isUseHcc = true;
+      this.isUseHcc = true;
       writeFileSync(join(distDir, 'midway_hcc.js'), code);
-    } catch (e) {
-      console.log('hooksBundle err:', e);
+    } catch {
+      //
     }
 
     let preloadCode = '';
@@ -721,7 +724,7 @@ export class PackagePlugin extends BasePlugin {
     preloadCode += [
       // for midway hooks preload
       'let hccModules = {modules: []};',
-      isUseHcc
+      this.isUseHcc
         ? `try {
         hccModules = require('./midway_hcc.js');
         console.log("hccModules", hccModules);
@@ -740,7 +743,10 @@ export class PackagePlugin extends BasePlugin {
     const configuratioCode = join(distDir, 'configuration.js');
     if (existsSync(configuratioCode)) {
       preloadCode += `
-      const configurationModule = require('./configuration.js');
+      let configurationModule = require('./configuration.js');
+      if (configurationModule && configurationModule.default) {
+        configurationModule = configurationModule.default;
+      }
       if(typeof configurationModule === 'object') {
         const className = Object.keys(configurationModule).find(key => {
           const cls = configurationModule[key];
@@ -757,6 +763,23 @@ export class PackagePlugin extends BasePlugin {
         exports.Configuration = configurationModule
       }
       `;
+    }
+
+    // TODO: fix ncc bug
+    try {
+      // eslint-disable-next-line
+      const midwayFaasMod = require.resolve('@midwayjs/faas/package.json', {
+        paths: [this.getCwd()],
+      });
+      const faasIndexPath = join(dirname(midwayFaasMod), './dist/index.js');
+      let faasIndexCode = await readFile(faasIndexPath, 'utf8');
+      faasIndexCode = faasIndexCode.replace(
+        /"\.\/configuration"/g,
+        '"@midwayjs/faas/dist/configuration"'
+      );
+      await writeFile(faasIndexPath, faasIndexCode);
+    } catch {
+      //
     }
 
     this.setStore(
@@ -1286,6 +1309,18 @@ export class PackagePlugin extends BasePlugin {
 
     await Promise.all(
       entryList.map(async entry => {
+        if (entry === 'midway.config.js' || entry === 'jest.config.js') {
+          return;
+        }
+        const entryPath = join(this.midwayBuildPath, entry);
+        if (this.isUseHcc) {
+          // 如果使用了 hcc 编译 midway hooks，在入口需要添加 hooks的编译文件，解决 midway.config.js 文件加载问题
+          const entryCode = await readFile(entryPath, 'utf8');
+          await writeFile(
+            entryPath,
+            'require("./dist/midway_hcc.js");' + entryCode
+          );
+        }
         const entryName = entry.slice(0, -3);
         await forkNode(
           nccCli,
@@ -1294,13 +1329,13 @@ export class PackagePlugin extends BasePlugin {
             cwd: this.midwayBuildPath,
           }
         );
-        await remove(join(this.midwayBuildPath, entry));
+        await remove(entryPath);
         await move(
           join(
             this.midwayBuildPath,
             'ncc_build_tmp/' + entryName + '/index.js'
           ),
-          join(this.midwayBuildPath, entry)
+          entryPath
         );
       })
     );