+++ packages/cli-plugin-package/src/utils.ts
@@ -117,9 +117,9 @@ export const analysisDecorator = async (cwd: string, currentFunc?) => {
   const midwayCoreMod = findNpmModule(cwd, '@midwayjs/core');
   const {
     ServerlessTriggerCollector,
-    // MidwayFrameworkService,
-    // MidwayServerlessFunctionService,
-    // prepareGlobalApplicationContext,
+    MidwayFrameworkService,
+    MidwayServerlessFunctionService,
+    prepareGlobalApplicationContext,
     WebRouterCollector,
   } = require(midwayCoreMod);
   let result;
@@ -128,30 +128,33 @@ export const analysisDecorator = async (cwd: string, currentFunc?) => {
     collector = new ServerlessTriggerCollector(cwd);
     result = await collector.getFunctionList();
   } else {
-    // const midwayDecoratorMod = findNpmModule(cwd, '@midwayjs/decorator');
-    // const {
-    //   CONFIGURATION_KEY,
-    //   listModule,
-    //   Types,
-    // } = require(midwayDecoratorMod);
-    // const applicationContext = prepareGlobalApplicationContext({
-    //   baseDir: cwd,
-    //   appDir: cwd,
-    // });
-    // await applicationContext.getAsync(MidwayFrameworkService, [
-    //   applicationContext,
-    // ]);
-    // const cycles = listModule(CONFIGURATION_KEY);
-    // for (const cycle of cycles) {
-    //   if (cycle.target && Types.isClass(cycle.target)) {
-    //     await applicationContext.getAsync(cycle.target);
-    //   }
-    // }
-    // const midwayServerlessFunctionService = applicationContext.get(
-    //   MidwayServerlessFunctionService
-    // );
-    // result = await midwayServerlessFunctionService.getFunctionList();
-    if (!result?.length) {
+    const pkg = join(midwayCoreMod, 'package.json');
+    const corePkgJson = JSON.parse(readFileSync(pkg, 'utf-8'));
+    if (corePkgJson?.version?.[0] === '3' && prepareGlobalApplicationContext) {
+      const midwayDecoratorMod = findNpmModule(cwd, '@midwayjs/decorator');
+      const {
+        CONFIGURATION_KEY,
+        listModule,
+        Types,
+      } = require(midwayDecoratorMod);
+      const applicationContext = prepareGlobalApplicationContext({
+        baseDir: cwd,
+        appDir: cwd,
+      });
+      await applicationContext.getAsync(MidwayFrameworkService, [
+        applicationContext,
+      ]);
+      const cycles = listModule(CONFIGURATION_KEY);
+      for (const cycle of cycles) {
+        if (cycle.target && Types.isClass(cycle.target)) {
+          await applicationContext.getAsync(cycle.target);
+        }
+      }
+      const midwayServerlessFunctionService = applicationContext.get(
+        MidwayServerlessFunctionService
+      );
+      result = await midwayServerlessFunctionService.getFunctionList();
+    } else {
       collector = new WebRouterCollector(cwd, {
         includeFunctionRouter: true,
       });